+++ src/app/auth/new-password/page.tsx
@@ -1,100 +1,44 @@
 'use client'
-import { useRouter, useSearchParams } from 'next/navigation'
+import { useSearchParams } from 'next/navigation'
 import { Button } from '@/shared/ui/base/Button/Button'
 import { Input } from '@/shared/ui/base/Input/Input'
-import { PATH } from '@/shared/lib/path/path'
-import { FormProvider, useForm } from 'react-hook-form'
-import { registrationSchema } from '@/shared/lib/schemas'
-import z from 'zod'
+import { FormProvider } from 'react-hook-form'
 import { Cards } from '@/shared/ui/base/Cards/Cards'
-import { zodResolver } from '@hookform/resolvers/zod'
-import { useCreateNewPasswordMutation } from '@/shared/api'
-import { useState } from 'react'
 import { Modal } from '@/shared/ui/Modal/Modal'
-
-const newPasswordSchema = registrationSchema.pick({
-  password: true,
-  confirmPassword: true,
-})
-
-type NewPasswordForm = z.infer<typeof newPasswordSchema>
+import { useCreateNewPassword } from '@/features/auth/new-password/model/useCreateNewPassword'
 
 export default function CreateNewPasswordPage() {
-  const router = useRouter()
-  const searchParams = useSearchParams()
-  const recoveryCode = searchParams.get('code') // код приходит в ссылке из письма
-  const email = searchParams.get('email') || ''
-  const [createNewPassword] = useCreateNewPasswordMutation()
+   const searchParams = useSearchParams()
+    const recoveryCode = searchParams.get('code')
+    const email = searchParams.get('email') || ''
 
-  const [modalOpen, setModalOpen] = useState(false)
-  const [modalMessage, setModalMessage] = useState('')
-  const [modalTitle, setModalTitle] = useState('')
+ const { methods, onSubmit, modal, showModal, closeModal } = useCreateNewPassword(
+        recoveryCode,
+        email
+    );
 
-  const methods = useForm<NewPasswordForm>({
-    resolver: zodResolver(newPasswordSchema),
-    defaultValues: {
-      password: '',
-      confirmPassword: '',
-    },
-    mode: 'onChange',
-  })
-  const {
-    handleSubmit,
-    formState: { isValid },
-  } = methods
-  const errors = methods.formState.errors
-
-  const showModal = (title: string, message: string) => {
-    setModalTitle(title)
-    setModalMessage(message)
-    setModalOpen(true)
-  }
-
-  const onSubmit = async (data: NewPasswordForm) => {
-    try {
-      await createNewPassword({
-        newPassword: data.password,
-        recoveryCode: recoveryCode || '',
-      }).unwrap()
-      showModal('Password successfully changed', 'success')
-      router.push(PATH.AUTH.LOGIN)
-    } catch (error: any) {
-      if (error.status === 400 && error.data?.error === 'Invalid or expired link') {
-        router.push(`${PATH.AUTH.RECOVERY_RESENDING}?email=${email}`)
-      } else if (error.status === 400) {
-        showModal('Incorrect data. Please try again.', 'error')
-      } else if (error.status === 429) {
-        showModal('Too many attempts. Please wait and try again.', 'error')
-      } else {
-        showModal('Server error', 'error')
-      }
-    }
-  }
+const {
+        formState: { isValid },
+    } = methods
 
   return (
-    <div className="flex items-center justify-center min-h-screen bg-[var(--color-dark-200)]">
-      <Modal open={modalOpen} onClose={() => setModalOpen(false)} modalTitle={modalTitle}>
-        {modalMessage}
+    <div className="flex items-center justify-center mt-3">
+      <Modal open={modal.open} onClose={closeModal} modalTitle={modal.title}>
       </Modal>
-      <div className="bg-[var(--color-dark-500)] p-8 shadow-lg w-[378px] h-[432px]">
-        <h2 className="text-xl font-semibold text-center mb-4 text-[var(--color-light-100)]">Create New Password</h2>
-
         <FormProvider {...methods}>
-          <Cards onSubmit={handleSubmit(onSubmit)}>
+          <Cards onSubmit={onSubmit}>
+              <h1 className="text-center text-h1">Create New Password</h1>
             <Input
               type="password"
               name="password"
               label="Enter new password"
-
-            />
-          
-
-            <Input
+ />
+         <Input
               type="password"
               name="confirmPassword"
               label="Confirm password"
             />
-  <p className="text-xs text-[var(--color-light-900)]">
+            <p className="text-xs ">
               Your password must be between 6 and 20 characters
             </p>
  <div className="mt-5 mb-5">
@@ -110,8 +54,56 @@ export default function CreateNewPasswordPage() {
           </Cards>
         </FormProvider>
       </div>
-    </div>
   )
 }
 
-////
+// const newPasswordSchema = registrationSchema.pick({
+//   password: true,
+//   confirmPassword: true,
+// })
+
+// type NewPasswordForm = z.infer<typeof newPasswordSchema>
+
+// export default function CreateNewPasswordPage() {
+  // const router = useRouter()
+  // const searchParams = useSearchParams()
+  // const recoveryCode = searchParams.get('code') 
+  // const email = searchParams.get('email') || ''
+  // const [createNewPassword] = useCreateNewPasswordMutation()
+  // const {modal, showModal, closeModal} = useModal();
+ 
+// const { methods, onSubmit } = useCreateNewPassword();
+  // const methods = useForm<NewPasswordForm>({
+  //   resolver: zodResolver(newPasswordSchema),
+  //   defaultValues: {
+  //     password: '',
+  //     confirmPassword: '',
+  //   },
+  //   mode: 'onChange',
+  // })
+  // const {
+  //   handleSubmit,
+  //   formState: { isValid },
+  // } = methods
+  // const errors = methods.formState.errors
+
+  // const onSubmit = async (data: NewPasswordForm) => {
+  //   try {
+  //     await createNewPassword({
+  //       newPassword: data.password,
+  //       recoveryCode: recoveryCode || '',
+  //     }).unwrap()
+  //     showModal('Password successfully changed', 'success')
+  //     router.push(PATH.AUTH.LOGIN)
+  //   } catch (error: any) {
+  //     if (error.status === 400 && error.data?.error === 'Invalid or expired link') {
+  //       router.push(`${PATH.AUTH.RECOVERY_RESENDING}?email=${email}`)
+  //     } else if (error.status === 400) {
+  //       showModal('Incorrect data. Please try again.', 'error')
+  //     } else if (error.status === 429) {
+  //       showModal('Too many attempts. Please wait and try again.', 'error')
+  //     } else {
+  //       showModal('Server error', 'error')
+  //     }
+  //   }
+  // }
