+++ src/app/auth/recovery-resending/page.tsx
@@ -5,21 +5,13 @@ import { useState } from "react";
 import { Button } from "@/shared/ui/base/Button/Button";
 import { Modal } from "@/shared/ui/Modal/Modal";
 import Image from "next/image";
+import { useModal } from "@/features/auth/forgot-passwors/model/useModal";
 
 const ResendLinkPage = () => {
   const searchParams = useSearchParams();
   const email = searchParams.get("email") || ""; 
   const [forgotPassword] = useForgotPasswordMutation();
-
-  const [modalOpen, setModalOpen] = useState(false);
-  const [modalMessage, setModalMessage] = useState("");
-  const [modalTitle, setModalTitle] = useState("");
-
-  const showModal = (title: string, message: string) => {
-    setModalTitle(title);
-    setModalMessage(message);
-    setModalOpen(true);
-  };
+   const {modal, showModal, closeModal} = useModal();
 
   const handleResend = async () => {
     try {
@@ -31,22 +23,20 @@ const ResendLinkPage = () => {
   };
 
   return (
-      <div className="flex items-center justify-center min-h-screen bg-[var(--color-dark-800)] px-4">
-      <div className="bg-[var(--color-dark-800)] rounded-[8px] flex flex-col items-center w-[378px] p-6">
-      <p className="text-h1 text-center text-[var(--color-light-100)] font-bold mb-5">
+    <div className="flex items-center justify-center px-4 rounded-[8px] flex-col w-[378px] p-6">
+      <p className="text-h1 text-center mb-5">
       Email verification link expired</p>
       <Modal
-        open={modalOpen}
-        onClose={() => setModalOpen(false)}
-        modalTitle={modalTitle}
+        open={modal.open}
+        onClose={closeModal}
+        modalTitle={modal.title}
       >
-        {modalMessage}
       </Modal>
-        <p className="text-regular_text14 text-center text-[var(--color-light-900)] mb-6">
+        <p className="text-xs ">
           Looks like the verification link has expired. <br />
           Not to worry, we can send the link again.
         </p>
-        <div className="mb-6">
+        <div className=" mt-6 mb-6">
         <Button
           type={"submit"}
           variant={"primary"}
@@ -58,8 +48,7 @@ const ResendLinkPage = () => {
         </Button>
         </div>
         <Image src="/rafiki.png" alt="rafiki" width={473} height={352} style={{ marginTop: "8px", objectFit: "contain" }}/>
-   </div>
-      </div>
+  </div>
   );
 };
 
@@ -67,27 +56,3 @@ export default ResendLinkPage;
 
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-// const Page = () => {
-//   return (
-//     <div>
-//       <button>Resend link</button>
-//     </div>
-//   )
-// }
-
-// export default Page
